Quite a few changes in this
Added a map page which is pretty much non-functional right now on either platform.  I did get all the packages in for maps, however, so it compiles.

It may throw an error when opening the Android map page (hitting the login button).

TODO - Figure out how to make map full screen in XAML.  Make a Controller for map page.  Fix permissions on Andriod and get API key for maps.