Made some major changes to the requester map page.
There is now a button on the bottom that will drop a dragable pin.  Theoretically, a user could use the pin or the search to find their destination at this point.

The app now checks where you are and centers the map on you if you are a requester.  It does this using a new NuGet: Xam.Plugin.Geolocator.

App is tested on both iOS and Android, however there is a known issue when location services are turned off.