DriverMapPage Changes:
Added an avgLat and Long to be used by a function that will resize the window.

Added a onPinClick method that currently only displays a pop-up.  Trying to get it to read the selected Pin in but I've been stumped on this for a while.

Added some logic to recenter map based on average lat and long of pins.  Doesn't guarentee all pins are on the map, but we may not want to force that anyways.

MapPage:
Added some logic so that the starting address is filled in automatically.  If the app can't get the user's location, it will just put "Starting Position..." like it does now.